✨ Add: 프로필 이미지관련 스토리지 로직 수정
- 기존 스토리지에 이미지를 올리면 이전 사용자의 이미지가 계속해서 쌓이는 경우에서 프로필 이미지가 업로드 되는 방식으로 변경

- Service 폴더 삭제, 해당 내용 AuthViewModel에서 작동하도록 변경
